src/requirements.txt
fastapi
uvicorn
pydantic
requests
PyPDF2
python-dotenv
python-multipart